Organize project components
